Split storage component into three classes.
