+++ src/main/java/de/craften/plugins/rpgplus/components/storage/MemoryStorageComponent.java
@@ -1,0 +1,14 @@
+package de.craften.plugins.rpgplus.components.storage;
+
+public class MemoryStorageComponent extends StorageComponent{
+
+	public MemoryStorageComponent() {
+		super(StorageType.MEMORY);
+	}
+	
+	@Override
+	protected void onActivated() {
+		storage = new MemoryStorage();
+	}
+	
+}
