+++ src/main/java/de/craften/plugins/rpgplus/components/storage/PdsStorageComponent.java
@@ -1,0 +1,14 @@
+package de.craften.plugins.rpgplus.components.storage;
+
+public class PdsStorageComponent extends StorageComponent{
+
+	public PdsStorageComponent() {
+		super(StorageType.PDS);
+	}
+	
+	@Override
+	protected void onActivated() {
+		storage = new PDSStorage();
+	}
+	
+}
